rewrite & update scientific calc feat
